# api/openapi.yaml
openapi: 3.0.3
info:
    description: Tenant-aware delivery and commerce API.
    title: Go Delivery App API
    version: 1.0.0
components:
    securitySchemes:
        # Authentication schemes
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http-tools
    schemas:
        ErrorResponse:
            properties:
                code:
                    type: string
                errors:
                    type: string
            type: object
...
        CreateAreaRequest:
            properties:
                city:
                    enum:
                        - Karachi
                        - Lahore
                    type: string
                name:
                    type: string
            required:
                - name
                - city
            type: object
        AreaResponse:
            properties:
                city:
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                name:
                    type: string
            type: object
        CreateZoneRequest:
            properties:
                coordinates_wkt:
                    type: string
                name:
                    type: string
            required:
                - name
                - coordinates_wkt
            type: object
        ZoneResponse:
            properties:
                area_id:
                    format: uuid
                    type: string
                coordinates_wkt:
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                name:
                    type: string
            type: object
        CreateMerchantServiceZoneRequest:
            properties:
                branch_id:
                    format: uuid
                    type: string
                zone_id:
                    format: uuid
                    type: string
            required:
                - zone_id
                - branch_id
            type: object
        MerchantServiceZoneResponse:
            properties:
                area_city:
                    type: string
                area_id:
                    format: uuid
                    type: string
                area_name:
                    type: string
                branch_id:
                    format: uuid
                    type: string
                branch_name:
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                zone_coordinates_wkt:
                    type: string
                zone_id:
                    format: uuid
                    type: string
                zone_name:
                    type: string
            type: object
        ServiceZoneCoverageCheckRequest:
            properties:
                latitude:
                    format: double
                    type: number
                longitude:
                    format: double
                    type: number
            required:
                - latitude
                - longitude
            type: object
        ServiceZoneCoverageCheckResponse:
            properties:
                area_city:
                    type: string
                area_id:
                    format: uuid
                    type: string
                area_name:
                    type: string
                branch_id:
                    format: uuid
                    type: string
                branch_name:
                    type: string
                covered:
                    type: boolean
                merchant_id:
                    format: uuid
                    type: string
                zone_id:
                    format: uuid
                    type: string
                zone_name:
                    type: string
            type: object
paths:
...
    /api/v1/areas:
        get:
            operationId: ListAreas
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/AreaResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: CreateArea
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAreaRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AreaResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/areas/{area_id}/zones:
        get:
            operationId: ListZonesByArea
            parameters:
                - in: path
                  name: area_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ZoneResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: CreateZone
            parameters:
                - in: path
                  name: area_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateZoneRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ZoneResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/merchant/service-zones:
        get:
            operationId: ListMerchantServiceZonesByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/MerchantServiceZoneResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
        post:
            operationId: CreateMerchantServiceZoneByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMerchantServiceZoneRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantServiceZoneResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/merchant/service-zones/check:
        post:
            operationId: CheckMerchantServiceZoneCoverage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ServiceZoneCoverageCheckRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ServiceZoneCoverageCheckResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error